# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TryNex Lifestyle API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
  productCount?: number;
}

export interface CreateCategoryRequest {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description?: string;
  price: number;
  discountPrice?: number;
  categoryId?: number;
  categoryName?: string;
  imageUrl?: string;
  images?: string[];
  sizes?: string[];
  colors?: string[];
  stock: number;
  featured?: boolean;
  rating?: number;
  reviewCount?: number;
  customizable?: boolean;
  tags?: string[];
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateProductRequest {
  name: string;
  slug: string;
  description?: string;
  price: number;
  discountPrice?: number;
  categoryId?: number;
  imageUrl?: string;
  images?: string[];
  sizes?: string[];
  colors?: string[];
  stock: number;
  featured?: boolean;
  customizable?: boolean;
  tags?: string[];
}

export interface UpdateProductRequest {
  name?: string;
  slug?: string;
  description?: string;
  price?: number;
  discountPrice?: number;
  categoryId?: number;
  imageUrl?: string;
  images?: string[];
  sizes?: string[];
  colors?: string[];
  stock?: number;
  featured?: boolean;
  customizable?: boolean;
  tags?: string[];
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productImage?: string;
  quantity: number;
  size?: string;
  color?: string;
  price: number;
  customNote?: string;
}

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cod: "cod",
  bkash: "bkash",
  nagad: "nagad",
  rocket: "rocket",
  card: "card",
} as const;

export type OrderPaymentStatus =
  (typeof OrderPaymentStatus)[keyof typeof OrderPaymentStatus];

export const OrderPaymentStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  ongoing: "ongoing",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  shippingAddress: string;
  shippingCity?: string;
  shippingDistrict?: string;
  paymentMethod: OrderPaymentMethod;
  paymentStatus?: OrderPaymentStatus;
  status: OrderStatus;
  items: OrderItem[];
  subtotal?: number;
  shippingCost?: number;
  total: number;
  notes?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CreateOrderRequestPaymentMethod =
  (typeof CreateOrderRequestPaymentMethod)[keyof typeof CreateOrderRequestPaymentMethod];

export const CreateOrderRequestPaymentMethod = {
  cod: "cod",
  bkash: "bkash",
  nagad: "nagad",
  rocket: "rocket",
  card: "card",
} as const;

export interface CreateOrderItemRequest {
  productId: number;
  quantity: number;
  size?: string;
  color?: string;
  customNote?: string;
}

export interface CreateOrderRequest {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  shippingAddress: string;
  shippingCity?: string;
  shippingDistrict?: string;
  paymentMethod: CreateOrderRequestPaymentMethod;
  items: CreateOrderItemRequest[];
  notes?: string;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  ongoing: "ongoing",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export interface TrackOrderRequest {
  orderNumber: string;
  email: string;
}

export interface SiteSettings {
  siteName: string;
  tagline?: string;
  phone?: string;
  email?: string;
  address?: string;
  facebookUrl?: string;
  instagramUrl?: string;
  heroTitle?: string;
  heroSubtitle?: string;
  announcementBar?: string;
  freeShippingThreshold?: number;
}

export interface UpdateSettingsRequest {
  siteName?: string;
  tagline?: string;
  phone?: string;
  email?: string;
  address?: string;
  facebookUrl?: string;
  instagramUrl?: string;
  heroTitle?: string;
  heroSubtitle?: string;
  announcementBar?: string;
  freeShippingThreshold?: number;
}

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  processingOrders: number;
  shippedOrders: number;
  deliveredOrders: number;
  totalRevenue: number;
  todayRevenue: number;
  totalProducts: number;
  lowStockProducts: number;
  recentOrders: Order[];
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token?: string;
}

export interface AdminMeResponse {
  authenticated: boolean;
  username?: string;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  featured?: boolean;
  page?: number;
  limit?: number;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  page?: number;
  limit?: number;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  ongoing: "ongoing",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;