# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TryNex Lifestyle API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders, admin]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, shipped, ongoing, delivered, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders, admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/track:
    post:
      operationId: trackOrder
      tags: [orders]
      summary: Track order by order number and email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackOrderRequest"
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Check admin auth status
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMeResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        quantity:
          type: integer
        size:
          type: string
        color:
          type: string
        price:
          type: number
        customNote:
          type: string
      required:
        - id
        - productId
        - productName
        - quantity
        - price
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: string
        shippingCity:
          type: string
        shippingDistrict:
          type: string
        paymentMethod:
          type: string
          enum: [cod, bkash, nagad, rocket, card]
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        status:
          type: string
          enum: [pending, processing, shipped, ongoing, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingCost:
          type: number
        total:
          type: number
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - customerName
        - customerEmail
        - customerPhone
        - shippingAddress
        - paymentMethod
        - status
        - items
        - total
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - limit
        - totalPages
    CreateOrderRequest:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: string
        shippingCity:
          type: string
        shippingDistrict:
          type: string
        paymentMethod:
          type: string
          enum: [cod, bkash, nagad, rocket, card]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemRequest"
        notes:
          type: string
      required:
        - customerName
        - customerEmail
        - customerPhone
        - shippingAddress
        - paymentMethod
        - items
    CreateOrderItemRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        size:
          type: string
        color:
          type: string
        customNote:
          type: string
      required:
        - productId
        - quantity
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, ongoing, delivered, cancelled]
      required:
        - status
    TrackOrderRequest:
      type: object
      properties:
        orderNumber:
          type: string
        email:
          type: string
      required:
        - orderNumber
        - email
...
    AdminLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
      required:
        - success
    AdminMeResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        username:
          type: string
      required:
        - authenticated