# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TryNex Lifestyle API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderItemRequest } from "./createOrderItemRequest";
import type { CreateOrderRequestPaymentMethod } from "./createOrderRequestPaymentMethod";

export interface CreateOrderRequest {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  shippingAddress: string;
  shippingCity?: string;
  shippingDistrict?: string;
  paymentMethod: CreateOrderRequestPaymentMethod;
  items: CreateOrderItemRequest[];
  notes?: string;
}