# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TryNex Lifestyle API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderPaymentStatus } from "./orderPaymentStatus";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  shippingAddress: string;
  shippingCity?: string;
  shippingDistrict?: string;
  paymentMethod: OrderPaymentMethod;
  paymentStatus?: OrderPaymentStatus;
  status: OrderStatus;
  items: OrderItem[];
  subtotal?: number;
  shippingCost?: number;
  total: number;
  notes?: string;
  createdAt?: Date;
  updatedAt?: Date;
}